add missing files

+++ include/linux/delay.h
@@ -56,7 +56,7 @@ static inline void ndelay(unsigned long x)
 extern unsigned long lpj_fine;
 void calibrate_delay(void);
 void __attribute__((weak)) calibration_delay_done(void);
-void msleep(unsigned long msecs);
+void msleep(unsigned int msecs);
 unsigned long msleep_interruptible(unsigned int msecs);
 void usleep_range(unsigned long min, unsigned long max);
